abractated core sprite functions to a base class
