+++ lib/compass/sass_extensions/functions/sprites.rb
@@ -1,5 +1,5 @@
 require 'digest/md5'
-
+require 'compass/sass_extensions/sprites/base'
 module Compass::SassExtensions::Functions::Sprites
   ZERO = Sass::Script::Number::new(0)
 
@@ -12,45 +12,7 @@ module Compass::SassExtensions::Functions::Sprites
     end
   end
 
-  class SpriteMap < Sass::Script::Literal
-
-    # Changing this string will invalidate all previously generated sprite images.
-    # We should do so only when the packing algorithm changes
-    SPRITE_VERSION = "1"
-
-    attr_accessor :image_names, :path, :name, :options
-    attr_accessor :images, :width, :height
-
-    def self.from_uri(uri, context, kwargs)
-      path, name = Compass::Sprites.path_and_name(uri.value)
-      sprites = Compass::Sprites.discover_sprites(uri.value).map do |sprite|
-        sprite.gsub(Compass.configuration.images_path+"/", "")
-      end
-      new(sprites, path, name, context, kwargs)
-    end
-
-    def initialize(image_names, path, name, context, options)
-      @image_names, @path, @name, @options = image_names, path, name, options
-      @images = nil
-      @width = nil
-      @height = nil
-      @evaluation_context = context
-      validate!
-      compute_image_metadata!
-    end
-
-    def sprite_names
-      image_names.map{|f| Compass::Sprites.sprite_name(f) }
-    end
-
-    def validate!
-      for sprite_name in sprite_names
-        unless sprite_name =~ /\A#{Sass::SCSS::RX::IDENT}\Z/
-          raise Sass::SyntaxError, "#{sprite_name} must be a legal css identifier"
-        end
-      end
-    end
-
+  class SpriteMap < Compass::SassExtensions::Sprites::Base
     # Calculates the overal image dimensions
     # collects image sizes and input parameters for each sprite
     def compute_image_metadata!
@@ -120,19 +82,6 @@ module Compass::SassExtensions::Functions::Sprites
       [width, height]
     end
 
-    # Generate a sprite image if necessary
-    def generate
-      if generation_required?
-        sprite_data = construct_sprite
-        save!(sprite_data)
-        Compass.configuration.run_callback(:sprite_generated, sprite_data)
-      end
-    end
-
-    def generation_required?
-      !File.exists?(filename) || outdated?
-    end
-
     def require_png_library!
       begin
         require 'oily_png'
@@ -160,11 +109,6 @@ module Compass::SassExtensions::Functions::Sprites
       output_png
     end
 
-    # The on-the-disk filename of the sprite
-    def filename
-      File.join(Compass.configuration.images_path, "#{path}-#{uniqueness_hash}.png")
-    end
-
     def uniqueness_hash
       @uniqueness_hash ||= begin
         sum = Digest::MD5.new
@@ -178,53 +122,6 @@ module Compass::SassExtensions::Functions::Sprites
         sum.hexdigest[0...10]
       end
       @uniqueness_hash
-    end
-
-    # saves the sprite for later retrieval
-    def save!(output_png)
-      saved = output_png.save filename
-      Compass.configuration.run_callback(:sprite_saved, filename)
-      saved
-    end
-
-    # All the full-path filenames involved in this sprite
-    def image_filenames
-      image_names.map do |image_name|
-        File.join(Compass.configuration.images_path, image_name)
-      end
-    end
-
-    # Checks whether this sprite is outdated
-    def outdated?
-      last_update = self.mtime
-      image_filenames.each do |image|
-        return true if File.mtime(image) > last_update
-      end
-      false
-    end
-
-    def mtime
-      File.mtime(filename)
-    end
-
-    def inspect
-      to_s
-    end
-
-    def to_s(options = self.options)
-      sprite_url(self).value
-    end
-
-    def respond_to?(meth)
-      super || @evaluation_context.respond_to?(meth)
-    end
-
-    def method_missing(meth, *args, &block)
-      if @evaluation_context.respond_to?(meth)
-        @evaluation_context.send(meth, *args, &block)
-      else
-        super
-      end
     end
 
   end
